4.15_lab_cloud_blogAPI/requirements.txt
azure-core==1.39.0
azure-cosmos==4.15.0
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
Flask==3.1.3
gunicorn==25.3.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.1
python-dotenv==1.2.2
requests==2.33.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
